refactor(network): move headers From/TryFrom to protobuf_messages and Iteration conversion to common
