+++ crates/papyrus_network/src/protobuf_messages/headers.rs
@@ -1,8 +1,3 @@
-use std::collections::HashMap;
-
-use futures::channel::mpsc::{Receiver, Sender};
-use futures::StreamExt;
-use prost::Message;
 use starknet_api::block::{
     BlockHash,
     BlockHeader,
@@ -27,15 +22,7 @@ use crate::protobuf_messages::{
     l1_data_availability_mode_to_enum_int,
     ProtobufConversionError,
 };
-use crate::{
-    BlockHashOrNumber,
-    Direction,
-    InternalQuery,
-    Protocol,
-    Query,
-    ResponseReceivers,
-    SignedBlockHeader,
-};
+use crate::{InternalQuery, Query, SignedBlockHeader};
 
 impl TryFrom<protobuf::BlockHeadersResponse> for Option<SignedBlockHeader> {
     type Error = ProtobufConversionError;
@@ -308,48 +295,17 @@ impl TryFrom<protobuf::BlockHeadersResponse> for Data {
 impl TryFrom<protobuf::BlockHeadersRequest> for InternalQuery {
     type Error = ProtobufConversionError;
     fn try_from(value: protobuf::BlockHeadersRequest) -> Result<Self, Self::Error> {
-        let value = value.iteration.ok_or(ProtobufConversionError::MissingField {
-            field_description: "BlockHeadersRequest::iteration",
-        })?;
-        let start = value.start.ok_or(ProtobufConversionError::MissingField {
-            field_description: "Iteration::start",
-        })?;
-        let start_block = match start {
-            protobuf::iteration::Start::BlockNumber(block_number) => {
-                BlockHashOrNumber::Number(BlockNumber(block_number))
-            }
-            protobuf::iteration::Start::Header(protobuf_hash) => {
-                BlockHashOrNumber::Hash(BlockHash(protobuf_hash.try_into()?))
-            }
-        };
-        let direction = match value.direction {
-            0 => Direction::Forward,
-            1 => Direction::Backward,
-            direction => {
-                return Err(ProtobufConversionError::OutOfRangeValue {
-                    type_description: "Direction",
-                    value_as_str: format!("{direction}"),
-                });
-            }
-        };
-        let limit = value.limit;
-        let step = value.step;
-        Ok(Self { start_block, direction, limit, step })
+        value
+            .iteration
+            .ok_or(ProtobufConversionError::MissingField {
+                field_description: "BlockHeadersRequest::iteration",
+            })?
+            .try_into()
     }
 }
 
 impl From<Query> for protobuf::BlockHeadersRequest {
     fn from(value: Query) -> Self {
-        protobuf::BlockHeadersRequest {
-            iteration: Some(protobuf::Iteration {
-                direction: match value.direction {
-                    Direction::Forward => 0,
-                    Direction::Backward => 1,
-                },
-                limit: value.limit as u64,
-                step: value.step as u64,
-                start: Some(protobuf::iteration::Start::BlockNumber(value.start_block.0)),
-            }),
-        }
+        protobuf::BlockHeadersRequest { iteration: Some(value.into()) }
     }
 }
